quant_ensemble/pyproject.toml
[tool.poetry]
name = "quant_ensemble"
version = "0.1.0"
description = "Production-ready ensemble quant trading system with Bloomberg integration"
authors = ["Quant Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
torch = "^2.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
tqdm = "^4.65.0"
loguru = "^0.7.0"
pytest = "^7.4.0"
blpapi = "^3.24.0"
scipy = "^1.11.0"
joblib = "^1.3.0"
pyarrow = "^14.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.0"
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
